write mutation length in commitlog as int, not long patch by jbellis; reviewed by gdusbabek for CASSANRA-1032
git-svn-id: https://svn.apache.org/repos/asf/cassandra/trunk@939059 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/cassandra/db/commitlog/CommitLog.java
@@ -203,7 +203,7 @@ public class CommitLog
                 byte[] bytes;
                 try
                 {
-                    bytes = new byte[(int) reader.readLong()]; // readlong can throw EOFException too
+                    bytes = new byte[reader.readInt()]; // readInt can throw EOFException too
                     reader.readFully(bytes);
                     claimedCRC32 = reader.readLong();
                 }
